# public/api-docs.yaml
openapi: 3.0.0
info:
  title: ScamGuard API
  description: Comprehensive scam detection and prevention API
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/admin/reports:
    get:
      tags: [Admin]
      summary: Get reports
      security:
        - BearerAuth: []
      responses:
        200:
          description: OK
    post:
      tags: [Admin]
      summary: Create report
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type: {type: string}
                content: {type: string}
      responses:
        200:
          description: OK
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: Get all users
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        200:
          description: OK
    post:
      tags: [Admin]
      summary: Create user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: {type: string}
                password: {type: string}
                name: {type: string}
      responses:
        200:
          description: OK
...
  /api/admin/settings:
    get:
      tags: [Admin]
      summary: Get settings
      security:
        - BearerAuth: []
      responses:
        200:
          description: OK
    put:
      tags: [Admin]
      summary: Update settings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        200:
          description: OK
  /api/admin/support:
    get:
      tags: [Admin]
      summary: Get support tickets
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        200:
          description: OK
    post:
      tags: [Admin]
      summary: Create support ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject, message]
              properties:
                subject: {type: string}
                message: {type: string}
                email: {type: string}
      responses:
        200:
          description: OK
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT